requirements.txt
Flask
python-dotenv
sentence_transformers
google-generativeai==0.6.0
IPython
flask-cors
mysql-connector-python
PyMySQL